SQL_Script_Procedure
"What you see Here?" This is a sample of a transformation from DWH tables that I developed to create an algorithm to forecast 5 main metrics of the Customer Partner and create goals for different times in a project of quarters. It is not need to check all the script, could be check some functions used to transform data and calculate values. (Is not been used  with 'x' as as postgreSQL permits)